+++ lib/rubocop/cop/style/redundant_freeze.rb
@@ -13,15 +13,16 @@ module RuboCop
       #   # good
       #   CONST = 1
       class RedundantFreeze < Cop
+        include FrozenStringLiteral
+
         MSG = 'Freezing immutable objects is pointless.'.freeze
 
         def on_send(node)
           receiver, method_name, *args = *node
 
-          return unless receiver &&
-                        method_name == :freeze &&
+          return unless method_name == :freeze &&
                         args.empty? &&
-                        receiver.immutable_literal?
+                        immutable_literal?(receiver)
 
           add_offense(node, :expression)
         end
@@ -31,6 +32,15 @@ module RuboCop
             corrector.remove(node.loc.dot)
             corrector.remove(node.loc.selector)
           end
+        end
+
+        private
+
+        def immutable_literal?(node)
+          return false unless node
+          return true if node.immutable_literal?
+          FROZEN_STRING_LITERAL_TYPES.include?(node.type) &&
+            frozen_string_literals_enabled?(processed_source)
         end
       end
     end